+++ src/inmanta/protocol/common.py
@@ -178,6 +178,7 @@ class ReturnValue(Generic[T]):
         headers: MutableMapping[str, str] = {},
         response: Optional[T] = None,
         content_type: str = JSON_CONTENT,
+        links: Optional[Dict[str, str]] = None,
     ) -> None:
         self._status_code = status_code
         self._warnings: List[str] = []
@@ -185,6 +186,7 @@ class ReturnValue(Generic[T]):
         self._headers[CONTENT_TYPE] = content_type
         self._content_type = content_type
         self._response = response
+        self._links = links
 
     @property
     def status_code(self) -> int:
@@ -216,6 +218,8 @@ class ReturnValue(Generic[T]):
 
         if len(self._warnings):
             response["metadata"] = {"warnings": self._warnings}
+        if self._links:
+            response["links"] = self._links
 
         return response
 
@@ -730,6 +734,8 @@ class MethodProperties(object):
         for i in range(len(args)):
             msg[argspec.args[i]] = args[i]
 
+        path_params = {k for k, v in msg.items() if v is not None and f"<{k}>" in self._path.path}
+
         url = self.get_call_url(msg)
 
         headers = {}
@@ -745,7 +751,7 @@ class MethodProperties(object):
                     del msg[arg_name]
 
         if self.operation not in ("POST", "PUT", "PATCH"):
-            qs_map = {k: v for k, v in msg.items() if v is not None and k != "id"}
+            qs_map = {k: v for k, v in msg.items() if v is not None and k not in path_params}
 
             # encode arguments in url
             if len(qs_map) > 0: